requirements.txt
absl-py==1.0.0
APScheduler==3.9.1
asgiref==3.5.2
astunparse==1.6.3
async-generator==1.10
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.5
cryptography==37.0.4
cycler==0.11.0
distlib==0.3.5
dj-database-url==1.0.0
Django==4.1
django-cors-headers==3.13.0
djangorestframework==3.13.1
et-xmlfile==1.1.0
filelock==3.7.1
flatbuffers==2.0
fonttools==4.30.0
gast==0.5.3
google-auth==2.6.0
google-auth-oauthlib==0.4.6
google-cloud-texttospeech==2.12.1
google-pasta==0.2.0
googletrans==4.0.0rc1
greenlet==1.1.2
grpcio==1.44.0
grpcio-status==1.47.0
gTTS==2.2.4
gunicorn==20.1.0
h11==0.9.0
h2==3.2.0
h5py==3.6.0
hpack==3.0.0
hstspreload==2022.8.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==3.3
importlib-metadata==4.11.2
install==1.3.5
itsdangerous==2.1.0
Jinja2==3.0.3
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.0
libclang==13.0.0
Markdown==3.3.6
MarkupSafe==2.1.0
matplotlib==3.5.1
nodejs==0.1.1
numpy==1.22.3
oauthlib==3.2.0
opencv-python-headless==4.5.5.64
openpyxl==3.0.10
opt-einsum==3.3.0
optional-django==0.1.0
outcome==1.2.0
packaging==21.3
pafy==0.5.5
Pillow==9.0.1
platformdirs==2.5.2
protobuf==3.19.4
psycopg2==2.9.3
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydub==0.25.1
pyOpenSSL==22.0.0
pyparsing==3.0.7
pyrubberband==0.3.0
PySocks==1.7.1
PySoundFile==0.9.0.post1
python-dateutil==2.8.2
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
requests-oauthlib==1.3.1
rfc3986==1.5.0
rsa==4.8
scipy==1.9.1
selenium==4.4.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
SoundFile==0.10.3.post1
SpeechRecognition==3.8.1
SQLAlchemy==1.4.31
sqlparse==0.4.2
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-cpu==2.8.0
tensorflow-io==0.26.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
tf-slim==1.1.0
trio==0.21.0
trio-websocket==0.9.2
typing-extensions==4.1.1
tzdata==2022.2
tzlocal==4.2
urllib3==1.26.8
virtualenv==20.16.2
Werkzeug==2.0.3
wrapt==1.14.0
wsproto==1.1.0
youtube-dl==2020.12.2
zipp==3.7.0
